# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AnalyzeRequestMode = typeof AnalyzeRequestMode[keyof typeof AnalyzeRequestMode];


export const AnalyzeRequestMode = {
  speech: 'speech',
  video: 'video',
} as const;

export interface AnalyzeRequest {
  sessionId: string;
  consentAccepted: boolean;
  mode: AnalyzeRequestMode;
  sourceName: string;
  sourceUri?: string;
  durationSeconds?: number;
  rawVideoPath?: string;
  extractedAudioPath?: string;
}

export interface Metric {
  score: number;
  label: string;
  detail: string;
}

export type AnalysisResultMode = typeof AnalysisResultMode[keyof typeof AnalysisResultMode];


export const AnalysisResultMode = {
  speech: 'speech',
  video: 'video',
} as const;

export interface AnalysisResult {
  id: string;
  createdAt: string;
  sourceName: string;
  mode: AnalysisResultMode;
  overallScore: number;
  confidence: Metric;
  eyeContact: Metric;
  speechRate: Metric;
  bodyPosture: Metric;
  handGestures: Metric;
  suggestions: string[];
}

export interface ErrorResponse {
  message: string;
}