# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analysis
    description: Speech and body language analysis
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /analyze:
    post:
      operationId: analyzePresentation
      tags: [analysis]
      summary: Analyze a speech or video presentation
      description: Accepts presentation metadata and returns coaching metrics.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Invalid analysis request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyze/{sessionId}:
    delete:
      operationId: deleteAnalysisSession
      tags: [analysis]
      summary: Delete all analysis data for a session
      description: Removes the analysis result and any retained raw media for a session.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Analysis data deleted
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AnalyzeRequest:
      type: object
      properties:
        sessionId:
          type: string
        consentAccepted:
          type: boolean
        mode:
          type: string
          enum: [speech, video]
        sourceName:
          type: string
        sourceUri:
          type: string
        durationSeconds:
          type: number
        rawVideoPath:
          type: string
        extractedAudioPath:
          type: string
      required:
        - sessionId
        - consentAccepted
        - mode
        - sourceName
    Metric:
      type: object
      properties:
        score:
          type: number
        label:
          type: string
        detail:
          type: string
      required:
        - score
        - label
        - detail
    AnalysisResult:
      type: object
      properties:
        sessionId:
          type: string
        id:
          type: string
        createdAt:
          type: string
        sourceName:
          type: string
        mode:
          type: string
          enum: [speech, video]
        overallScore:
          type: number
        confidence:
          $ref: "#/components/schemas/Metric"
        eyeContact:
          $ref: "#/components/schemas/Metric"
        speechRate:
          $ref: "#/components/schemas/Metric"
        bodyPosture:
          $ref: "#/components/schemas/Metric"
        handGestures:
          $ref: "#/components/schemas/Metric"
        suggestions:
          type: array
          items:
            type: string
      required:
        - sessionId
        - id
        - createdAt
        - sourceName
        - mode
        - overallScore
        - confidence
        - eyeContact
        - speechRate
        - bodyPosture
        - handGestures
        - suggestions
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message