# lib/api-zod/src/generated/types/analyzeRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeRequestMode } from './analyzeRequestMode';

export interface AnalyzeRequest {
  sessionId: string;
  consentAccepted: boolean;
  mode: AnalyzeRequestMode;
  sourceName: string;
  sourceUri?: string;
  durationSeconds?: number;
  rawVideoPath?: string;
  extractedAudioPath?: string;
}